requirements.txt
click
pyyaml
pytest
